JavaDoc some Util methods and organize exceptions
