+++ src/main/java/com/dsh105/nexus/command/module/CICommand.java
@@ -21,8 +21,7 @@ import com.dsh105.nexus.Nexus;
 import com.dsh105.nexus.command.Command;
 import com.dsh105.nexus.command.CommandModule;
 import com.dsh105.nexus.command.CommandPerformEvent;
-import com.dsh105.nexus.exception.JenkinsException;
-import com.dsh105.nexus.exception.JenkinsJobNotFoundException;
+import com.dsh105.nexus.exception.jenkins.JenkinsException;
 import com.dsh105.nexus.hook.jenkins.JenkinsJob;
 import com.dsh105.nexus.hook.jenkins.Result;
 import com.dsh105.nexus.util.StringUtil;
