+++ src/main/java/com/dsh105/nexus/exception/general/GenericUrlConnectionException.java
@@ -15,15 +15,11 @@
  * along with Nexus.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dsh105.nexus.exception;
+package com.dsh105.nexus.exception.general;
 
-public class JenkinsJobNotFoundException extends JenkinsException {
+public class GenericUrlConnectionException extends RuntimeException {
 
-    public JenkinsJobNotFoundException(String s) {
-        super(s);
-    }
-
-    public JenkinsJobNotFoundException(String s, Throwable throwable) {
+    public GenericUrlConnectionException(String s, Throwable throwable) {
         super(s, throwable);
     }
 }
