+++ src/main/java/com/dsh105/nexus/exception/github/GitHubRateLimitExceededException.java
@@ -15,11 +15,11 @@
  * along with Nexus.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dsh105.nexus.exception;
+package com.dsh105.nexus.exception.github;
 
-public class GitHubHookNotFoundException extends RuntimeException {
+public class GitHubRateLimitExceededException extends GitHubException {
 
-    public GitHubHookNotFoundException(String s) {
+    public GitHubRateLimitExceededException(String s) {
         super(s);
     }
 }
